Add files via upload
This is a Windows compiled version of the same .bat file mentioned in https://github.com/rstudio/rstudio/issues/12463#issuecomment-1482326544
As @brry mentioned: "Note for others: the symbol in the file browser then is no longer the Rstudio logo.", this .exe file contains the RStudio icon.
So, instead of linking your R related file to the .bat file (which does not have a custom icon), you can refer to rstudiomax.exe to open the R related files. 
This way the RStudio icon will be visible for all R related files.